Refactor aria translations
