+++ packages/react/src/alertbox/alertbox.tsx
@@ -10,7 +10,8 @@ import {
   Warning,
 } from '@obosbbl/grunnmuren-icons-react';
 import { useState } from 'react';
-import { type Locale, useLocale } from '../use-locale';
+import { useLocale } from '../use-locale';
+import { translations } from '../translations';
 
 const iconMap = {
   info: InfoCircle,
@@ -77,32 +78,6 @@ type Props = VariantProps<typeof alertVariants> & {
    * This is used to control the open/closed state of the component; make the component "controlled".
    */
   onDismiss?: () => void;
-};
-
-type Translation = {
-  [key in Locale]: string;
-};
-
-type Translations = {
-  [x: string]: Translation;
-};
-
-const translations: Translations = {
-  close: {
-    nb: 'Lukk',
-    sv: 'Stäng',
-    en: 'Close',
-  },
-  showMore: {
-    nb: 'Les mer',
-    sv: 'Läs mer',
-    en: 'Read more',
-  },
-  showLess: {
-    nb: 'Vis mindre',
-    sv: 'Dölj',
-    en: 'Show less',
-  },
 };
 
 const Alertbox = ({
